fix:read_heredoc.cの分割
